refactored workspace structure
