+++ src/plugin/accessories/CameraAccessory.ts
@@ -8,9 +8,9 @@ import { DeviceAccessory } from './Device';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore  
 import { Camera, Device, PropertyName, CommandName, VideoCodec } from 'eufy-security-client';
-import { StreamingDelegate } from './streamingDelegate';
+import { StreamingDelegate } from '../controller/streamingDelegate';
 
-import { CameraConfig, VideoConfig } from './configTypes';
+import { CameraConfig, VideoConfig } from '../utils/configTypes';
 
 /**
  * Platform Accessory
