+++ src/plugin/utils/configTypes.ts
